% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{correlate_cols}
\alias{correlate_cols}
\title{Pairwise correlation of the columns of two matrices}
\usage{
correlate_cols(data.1, data.2, metric = "pearson")
}
\arguments{
\item{data.1}{First input matrix}

\item{data.2}{Second input matrix}

\item{metric}{Metric to use (either pearson or cosine)}
}
\value{
Matrix of column to column correlations
}
\description{
Pairwise correlation of the columns of two matrices
}
