% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{CrossValidateKNN}
\alias{CrossValidateKNN}
\title{Cross validation to find the optimal k}
\usage{
CrossValidateKNN(train.knn, k.range, n.folds = 4, seed = NULL,
  n.cores = 4)
}
\arguments{
\item{train.knn}{List output of TrainKNN with gene expression matrix, PCs, and reference ident}

\item{k.range}{Range of k values to test}

\item{n.folds}{Number of folds to split the data}

\item{seed}{Random seed for splitting folds}

\item{n.cores}{Number of multithreading cores}
}
\value{
List of k value accuracies and the optimal k
}
\description{
Cross validation to find the optimal k
}
