% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{fast_pca}
\alias{fast_pca}
\title{Run PCA}
\usage{
fast_pca(x, pcs.use)
}
\arguments{
\item{x}{Input data matrix, should be genes x samples}

\item{pcs.use}{Number of PCs to calculate}
}
\value{
List of PCA embeddings, loadings, and variance explained
}
\description{
Run PCA
}
