% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{FastPCA}
\alias{FastPCA}
\title{Run PCA}
\usage{
FastPCA(norm.data, genes.use, pcs.use = 40)
}
\arguments{
\item{norm.data}{Input data matrix, should be genes x samples}

\item{genes.use}{Subset of genes to use}

\item{pcs.use}{Number of PCs to calculate}
}
\value{
List of PCA embeddings, loadings, and variance explained
}
\description{
Run PCA
}
