% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{MapCellsKNN}
\alias{MapCellsKNN}
\title{Map cells to a reference dataset using a kNN classifier}
\usage{
MapCellsKNN(query.data, ref.data, ref.ident, k = 30)
}
\arguments{
\item{query.data}{Query matrix}

\item{ref.data}{Reference matrix}

\item{ref.ident}{Reference cell types}

\item{k}{Number of neighbors}
}
\value{
Matrix of cell type probabilities
}
\description{
Map cells to a reference dataset using a kNN classifier
}
