% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{MapClustersCor}
\alias{MapClustersCor}
\title{Maps query clusters to reference clusters using correlation}
\usage{
MapClustersCor(query.data, query.clusters, train.data, train.clusters,
  genes.use = NULL, metric = "pearson")
}
\arguments{
\item{query.data}{Query gene expression matrix}

\item{query.clusters}{Factor of query clusters}

\item{train.data}{Reference gene expression matrix}

\item{train.clusters}{Factor of reference clusters}

\item{genes.use}{Genes to use for classification}

\item{metric}{Metric to use (either pearson or cosine)}
}
\value{
Matrix of query to reference cluster correlations
}
\description{
Maps query clusters to reference clusters using correlation
}
