% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{MapKNN}
\alias{MapKNN}
\title{Map cells to a reference dataset using a kNN classifier}
\usage{
MapKNN(query.norm.counts, train.knn, genes.use = NULL, use.pca = T,
  k = 30)
}
\arguments{
\item{query.norm.counts}{Query gene expression matrix}

\item{train.knn}{Output of TrainKNN}

\item{genes.use}{Genes to use for classification}

\item{use.pca}{Whether or not to project to PCs first}

\item{k}{Number of neighbors}
}
\value{
Matrix of cell type probabilities
}
\description{
Map cells to a reference dataset using a kNN classifier
}
