% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse.human.table}
\alias{mouse.human.table}
\title{Mouse to human gene name conversion set}
\format{A data frame with 27779 rows and 4 variables:}
\usage{
mouse.human.table
}
\description{
Mouse to human gene name conversion set
}
\keyword{datasets}
