% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{MouseHumanMapping}
\alias{MouseHumanMapping}
\title{Map mouse genes to human genes}
\usage{
MouseHumanMapping(mouse.genes)
}
\arguments{
\item{mouse.genes}{Vector of mouse genes}
}
\value{
vector of human genes with mouse genes as names
}
\description{
Map mouse genes to human genes
}
