% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{project_pca}
\alias{project_pca}
\title{Projects new data onto existing PCs}
\usage{
project_pca(newx, pc.load)
}
\arguments{
\item{newx}{Dataset to project, should be genes x samples}

\item{pc.load}{Existing pca loadings}
}
\value{
Projected PCA embeddings
}
\description{
Projects new data onto existing PCs
}
