% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{ProjectPCA}
\alias{ProjectPCA}
\title{Projects new data onto existing PCs}
\usage{
ProjectPCA(new.data, pc.load, genes.use)
}
\arguments{
\item{new.data}{Dataset to project, should be genes x samples}

\item{pc.load}{Existing pca loadings}

\item{genes.use}{Genes to use for projection}
}
\value{
Projected PCA embeddings
}
\description{
Projects new data onto existing PCs
}
