% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{SummarizeCellMapping}
\alias{SummarizeCellMapping}
\title{Summarize cell classification by group}
\usage{
SummarizeCellMapping(cell.ident.frac, query.ident, ident.cell.thresh = 0,
  hard.cell.thresh = T, min.ref.frac = 0.05)
}
\arguments{
\item{cell.ident.frac}{A reference groups x cells matrix of reference group probabilities}

\item{query.ident}{Either a list or a factor of cell type identities for the query dataset}

\item{ident.cell.thresh}{Filter out cells that don't map to a reference group with at least this probability}

\item{hard.cell.thresh}{Assigns each cell to a single reference group}

\item{min.ref.frac}{Filter out reference groups that are not mapped to with at least this probability}
}
\value{
A reference groups x query groups matrix of probabilities
}
\description{
Summarize cell classification by group
}
