% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{SummarizeKNN}
\alias{SummarizeKNN}
\title{Summarize cell classification by group}
\usage{
SummarizeKNN(cell.mapped.ident, query.ident, ident.thresh = 0,
  assign.cells = T, min.ident.frac = 0.05)
}
\arguments{
\item{cell.mapped.ident}{Mapped cell type probabilities}

\item{query.ident}{Either a list or a factor of cell type identities for the query dataset}

\item{ident.thresh}{Filter out cells that don't map to a reference identity}

\item{assign.cells}{Assigns each cell to a single reference group}

\item{min.ident.frac}{Filter out reference identities}
}
\value{
A reference groups x query groups matrix of probabilities
}
\description{
Summarize cell classification by group
}
