% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{TrainKNN}
\alias{TrainKNN}
\title{Runs PCA on the training dataset}
\usage{
TrainKNN(norm.counts, genes.use, ident, pcs.use = 40)
}
\arguments{
\item{norm.counts}{Normalized gene expression matrix, should be genes x samples}

\item{genes.use}{Genes to use for PCA}

\item{ident}{Cell type identity}

\item{pcs.use}{Number of PCs to calculate}
}
\value{
List of gene expression, cell identity, PCA embeddings, PCA loadings
}
\description{
Runs PCA on the training dataset
}
