% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{weighted_neighbor_voting}
\alias{weighted_neighbor_voting}
\title{Compute distance weighted voting scores for each reference class for each cell}
\usage{
weighted_neighbor_voting(knn.res, ident)
}
\arguments{
\item{knn.res}{List of neighbors and distances, list output of FNN::get.knnx}

\item{ident}{Reference identity}
}
\value{
Reference classes by cells matrix of voting scores
}
\description{
Compute distance weighted voting scores for each reference class for each cell
}
