% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{MapClustersCor}
\alias{MapClustersCor}
\title{Maps query clusters to reference clusters using correlation}
\usage{
MapClustersCor(query.norm.counts, query.clusters, train.knn,
  genes.use = NULL, use.pca = T, metric = "pearson")
}
\arguments{
\item{query.norm.counts}{Query gene expression matrix}

\item{query.clusters}{Factor of query clusters}

\item{train.knn}{Output of TrainKNN}

\item{genes.use}{Genes to use for classification}

\item{use.pca}{Whether or not to project to PCs first}

\item{metric}{Metric to use (either pearson or cosine)}
}
\value{
Matrix of query to reference cluster correlations
}
\description{
Maps query clusters to reference clusters using correlation
}
